Move test folder to extension's root
