+++ package.json
@@ -33,7 +33,7 @@
     "activationEvents": [
         "*"
     ],
-    "main": "./out/extension",
+    "main": "./out/src/extension",
     "categories": [
         "Other"
     ],
